+++ app/src/main/java/fi/dias/tools/vero/tasks/FetchCertificate.java
@@ -1,5 +1,9 @@
 package fi.dias.tools.vero.tasks;
 
+import com.evanlennick.retry4j.CallExecutorBuilder;
+import com.evanlennick.retry4j.Status;
+import com.evanlennick.retry4j.config.RetryConfig;
+import com.evanlennick.retry4j.config.RetryConfigBuilder;
 import fi.vero.certificates._2017._10.certificateservices.*;
 
 import java.io.ByteArrayInputStream;
@@ -7,8 +11,10 @@ import java.io.InputStream;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
+import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.Callable;
 import java.util.logging.Logger;
 
 public class FetchCertificate extends WebServiceTask{
@@ -59,14 +65,29 @@ public class FetchCertificate extends WebServiceTask{
         if (this.prepare()) {
             throw new UnsupportedOperationException("Trying to fetch certificate while missing necessary information, errors: " + String.join(", ", errors));
         }
-
-        GetCertificateResponse getCertificateResponse = this.port.getCertificate(this.getCertificateRequest);
-        Result result = getCertificateResponse.getResult();
-
-        this.handleErrorResult(result, "Failed to fetch certificate with retrieval id:" + this.retrievalId);
-
         CertificateFactory certificateFactory = CertificateFactory.getInstance("X.509");
-        InputStream in = new ByteArrayInputStream(getCertificateResponse.getCertificate());
-        return (X509Certificate) certificateFactory.generateCertificate(in);
+
+        Callable<ByteArrayInputStream> callable = () -> {
+            GetCertificateResponse getCertificateResponse = this.port.getCertificate(this.getCertificateRequest);
+            Result result = getCertificateResponse.getResult();
+
+            this.handleErrorResult(result, "Failed to fetch certificate with retrieval id:" + this.retrievalId);
+            return new ByteArrayInputStream(getCertificateResponse.getCertificate());
+        };
+
+        RetryConfig config = new RetryConfigBuilder()
+                .retryOnAnyException()
+                .withFixedBackoff()
+                .withDelayBetweenTries(45, ChronoUnit.SECONDS)
+                .withMaxNumberOfTries(10)
+                .build();
+
+        Status<ByteArrayInputStream> status = new CallExecutorBuilder().config(config).build().execute(callable);
+
+        if (status.wasSuccessful()) {
+            return (X509Certificate) certificateFactory.generateCertificate(status.getResult());
+        } else {
+            throw new RuntimeException("Failed to fetch certificate after 5 tries in 5 second");
+        }
     }
 }